+++ src/main/clojure/convex_web/web_server.clj
@@ -116,6 +116,27 @@
        ;; -- End Google Analytics
 
 
+       ;; -- HubSpot
+
+       [:script
+        {:id "hs-script-loader"
+         :async "true"
+         :defer "true"
+         :src "//js-na1.hs-scripts.com/24109496.js"}]
+
+       [:script
+        {:src "//js.hsforms.net/forms/embed/v2.js"}]
+
+       [:script
+        "hbspt.forms.create({
+            region: 'na1',
+            portalId: '24109496',
+            formId: 'bc3f0027-bc36-41d6-bfdb-c19700419d20'
+        });"]
+
+       ;; -- End HubSpot
+
+
        [:link
         {:rel "preconnect"
          :href "https://fonts.googleapis.com"}]
@@ -139,6 +160,8 @@
        [:title "Convex"]
 
        [:body
+        (page/include-js (str asset-prefix-url "/linkedin.js"))
+
         (ring.util.anti-forgery/anti-forgery-field)
 
         [:div#app]