+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Major changes and resolved issues
 
+## 2023-04-24
+- Setup LinkedIn Insight Tag
+
+## 2023-04-15
+- More HubSpot
+
+## 2023-04-14
+- Install HubSpot
+
 ## 2023-03-17
 - New page:
 	- Trademark Policy